+++ src/index.mjs
@@ -16,7 +16,7 @@ import { whichPMRuns } from 'which-pm-runs';
 const DEFAULT_CONFIG_FILE = 'stylelint.config.mjs';
 const DEFAULT_CONFIG_CONTENT = `/** @type {import("stylelint").Config} */
 export default {
-  "extends": ["stylelint-config-standard"]
+  extends: ["stylelint-config-standard"]
 };`;
 
 const ADD_COMMAND = 'add -D stylelint stylelint-config-standard';